+++ vet-clinic/src/main/java/com/vetclinic/blo0021/repository/HostRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HostRepository extends JpaRepository<Host, Long> {
+public interface HostRepository extends JpaRepository<Host, Integer> {
 }
