+++ vet-clinic/src/main/java/com/vetclinic/blo0021/repository/IllnessRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IllnessRepository extends JpaRepository<Illness, Long> {
+public interface IllnessRepository extends JpaRepository<Illness, Integer> {
 }
